% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-results.R
\name{collect_predictions}
\alias{collect_predictions}
\alias{calculate_mse}
\title{Download and evaluate prediction}
\usage{
collect_predictions(url)

calculate_mse(prediction, target, bootstrap_indices)
}
\arguments{
\item{url}{url to prediction file}

\item{target}{vector of truth}
}
\value{
parsed predictions
}
\description{
Download and evaluate prediction
}
\section{Functions}{
\itemize{
\item \code{calculate_mse}: calculate MSE and run a bootstrap
}}

