% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{conc_plot}
\alias{conc_plot}
\title{Create a scatterplot}
\usage{
conc_plot(data, .xVar, .yVar)
}
\arguments{
\item{data}{the data to use for the plot}

\item{.xVar}{the value to plot on the x axis}

\item{.yVar}{the value to plot on the y axis}
}
\description{
Create a scatterplot
}
\examples{
conc_plot(Theoph, Time, conc)
}
