% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-results.R
\name{download_predictions}
\alias{download_predictions}
\title{Download predictions}
\usage{
download_predictions(url)
}
\arguments{
\item{url}{url to prediction file}
}
\value{
parsed predictions
}
\description{
Download predictions
}
\examples{
download_predictions('https://raw.githubusercontent.com/Rochester-Biomedical-DS/Hackathon-Summer-2020/master/prediction/prediction.csv')
}
