% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-results.R
\name{last_modified}
\alias{last_modified}
\alias{has_readme}
\alias{eval_predictions}
\title{Download and evaluate prediction}
\usage{
last_modified(url)

has_readme(url)

eval_predictions(url, target)
}
\arguments{
\item{url}{url to prediction file}

\item{target}{vector of truth}
}
\value{
parsed predictions
}
\description{
Download and evaluate prediction
}
\section{Functions}{
\itemize{
\item \code{last_modified}: get last-modified header

\item \code{has_readme}: check for README.md
}}

