% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hackathon-methods.R
\name{get_local_or_remote_csv}
\alias{get_local_or_remote_csv}
\alias{get_local_or_remote_tsv}
\title{Load a csv from remote or local}
\usage{
get_local_or_remote_csv(row)

get_local_or_remote_tsv(row)
}
\arguments{
\item{row}{list (one row) from the MANIFEST.  \code{url} needs to be set.}
}
\value{
\code{tibble}
}
\description{
Load a csv from remote or local
}
\section{Functions}{
\itemize{
\item \code{get_local_or_remote_tsv}: load tsv from remote or local
}}

