% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackathon-package.R
\docType{package}
\name{hackathon-package}
\alias{hackathon}
\alias{hackathon-package}
\title{hackathon: Package Test}
\description{
Contient la fonction MakeTop.
}
\author{
\strong{Maintainer}: Loubna El Madouri \email{lou.e.m.lem@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
