% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hackathon-class.R
\name{Hackathon}
\alias{Hackathon}
\title{Construct a Hackathon object}
\usage{
Hackathon(parms, competitor_tbl)
}
\arguments{
\item{parms}{list of parameters for the hackathon}

\item{competitor_tbl}{data.frame/tibble of competitors}
}
\value{
S3 object
}
\description{
This encapulates a directory structure containing configuration and private data.
Various methods support scoring and updating the Hackathon.
}
\seealso{
setup_hack

load_hack
}
