% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hackathon-methods.R
\name{score_hack}
\alias{score_hack}
\title{Score a Hackathon by crawling a manifest}
\usage{
score_hack(
  hack,
  get_prediction_by_manifest = get_local_or_remote_csv,
  score_prediction,
  truth = hack[["truth"]],
  ...
)
}
\arguments{
\item{hack}{\code{Hackathon}}

\item{get_prediction_by_manifest}{a function operating on a row of the manifest, returning an object that will be scored by \code{score_prediction}}

\item{score_prediction}{a function operating on a loaded prediction and \code{truth}.  Returns a \code{data.frame}.}

\item{truth}{}

\item{...}{passed to \code{score_prediction}}
}
\value{
\code{tibble}
}
\description{
For each row in the manifest, read the prediction via \code{get_prediction_by_manifest}
and score it via \code{score_prediction}.
Both the reader and scorer function are wrapped in \code{purrr::safely} to trap errors,
which are added as fields to the score table
}
