% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hackathon-methods.R
\name{setup_hack}
\alias{setup_hack}
\alias{load_hack}
\title{Initialize a Hackathon directory and YAML}
\usage{
setup_hack(parent_path = ".", name, desc)

load_hack(hack_root = ".", ...)
}
\arguments{
\item{parent_path}{path (relative to working directory) to write Hackathon project}

\item{name}{Name of hackathon}

\item{desc}{Show description}

\item{hack_root}{path containing competition.yml file}
}
\value{
\link{Hackathon}
}
\description{
Initialize a Hackathon directory and YAML
}
\section{Functions}{
\itemize{
\item \code{load_hack}: load existing hackathon
}}

\examples{
h = setup_hack('.', 'Example', 'Test')
h
}
