% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{shinydata}
\alias{shinydata}
\title{A function that creates a scatterplot}
\usage{
shinydata(dataset)
}
\arguments{
\item{dataset}{the dataset to plot}
}
\description{
A function that creates a scatterplot
}
