% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hackathon-class.R
\name{Hackathon}
\alias{Hackathon}
\title{Construct a Hackathon object}
\usage{
Hackathon(parms, competitor_tbl, truth)
}
\arguments{
\item{parms}{list of parameters for the hackathon}

\item{competitor_tbl}{data.frame/tibble of competitors}

\item{truth}{data.frame/tibble of secret truth data}
}
\value{
S3 object
}
\description{
Note: you probably will want to initialize a Hackthon using \code{\link[=setup_hack]{setup_hack()}}.
This encapsulates a directory structure containing configuration and private data.
Various methods support scoring and updating the Hackathon.
}
\seealso{
setup_hack

load_hack
}
