% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{update_gs_manifest}
\alias{update_gs_manifest}
\title{Downlaod a manifest from googlesheets, add urls and save}
\usage{
update_gs_manifest(subsheet, competition_name)
}
\arguments{
\item{competition_name}{the name of the competition.
We will look for a repos with this name under the registered github handles.}

\item{gslink}{link to googlesheet.
You may need to authorize the googlesheets4 API or ortherwise get an API key}
}
\value{
\code{logical} if the manifest is OK.
}
\description{
Downlaod a manifest from googlesheets, add urls and save
}
